+++ packages/core/src/http-client/http-client-types.ts
@@ -62,6 +62,7 @@ interface KnownHttpResponseFields {
   data: any;
   status: number;
   headers: any;
+  request: any;
 }
 
 export interface HttpResponse extends KnownHttpResponseFields {
@@ -73,9 +74,4 @@ export interface HttpResponse extends KnownHttpResponseFields {
  */
 export interface HttpReponse extends KnownHttpResponseFields {
   [otherKey: string]: any;
-}
-
-export interface HttpRequestAndResponse{
-  request: HttpRequest;
-  response: HttpResponse;
 }