+++ packages/generator/src/generator-cli.spec.ts
@@ -1,10 +1,11 @@
 import * as path from 'path';
 import execa = require('execa');
 import * as fs from 'fs-extra';
+import { oDataServiceSpecs } from '../../../test-resources/odata-service-specs';
 
 describe('generator-cli', () => {
   const pathToGenerator = path.resolve(__dirname, 'generator-cli.ts');
-  const inputDir = '../../test-resources/odata-service-specs/v2/API_TEST_SRV/';
+  const inputDir = path.resolve(oDataServiceSpecs,'v2','API_TEST_SRV');
   const outputDir = path.resolve(__dirname, '../test/generator-test-output');
 
   beforeEach(() => {