+++ packages/generator/src/sdk-metadata/code-samples/generic-get-all-code-sample.ts
@@ -1,3 +1,11 @@
-import { BusinessPartner } from '@sap/cloud-sdk-vdm-business-partner-service';
+import { codeBlock } from '@sap-cloud-sdk/util';
 
-const resultPromise = BusinessPartner.requestBuilder().getAll().top(5).execute({ destinationName:'myDestinationName' });
+export function genericGetAllCodeSample(entityName: string,packageName: string): string{
+return codeBlock`
+import { ${entityName} } from '${packageName}';
+
+const resultPromise = ${entityName}.requestBuilder().getAll().top(5).execute({ destinationName:'myDestinationName' });
+
+`;
+}
+