+++ test-packages/type-tests/test/http-client.spec.ts
@@ -1,10 +1,6 @@
 import {
-  executeHttpRequest,
-  executeHttpRequestReturnRequestAndResponse
+  executeHttpRequest
 } from '@sap-cloud-sdk/core';
 
 // $ExpectType Promise<HttpResponse>
 executeHttpRequest({ url: 'https://example.com' }, { method: 'get' });
-
-// $ExpectType Promise<HttpRequestAndResponse>
-executeHttpRequestReturnRequestAndResponse({ url: 'https://example.com' }, { method: 'get' });